# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentación de API ",
    "version": "1.0.0",
    "title": "API BIBLIOTECA UNIVALLE",
    "contact": {
      "email": "vqg2021079@est.univalle.edu"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost",
  "basePath": "/univalle/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/books": {
      "get": {
        "tags": [
          "book"
        ],
        "summary": "Listado de libros",
        "description": "",
        "operationId": "listBook",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "book"
        ],
        "summary": "Add a new book in the register",
        "description": "",
        "operationId": "addBook",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "book object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Student created"
          }
        }
      },
      "put": {
        "tags": [
          "book"
        ],
        "summary": "Update an existing book",
        "description": "",
        "operationId": "updateBook",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Pet object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/book/{id}": {
      "get": {
        "tags": [
          "book"
        ],
        "summary": "Find student by ID",
        "description": "Returns a single book",
        "operationId": "getBookbyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of book to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "patch": {
        "tags": [
          "book"
        ],
        "summary": "Updates a book in the register with form data",
        "description": "",
        "operationId": "updateBooktWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the pet",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the pet",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "book"
        ],
        "summary": "Deletes a Book",
        "description": "",
        "operationId": "deleteBook",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    },
...
    "/author": {
      "get": {
        "tags": [
          "Author"
        ],
        "summary": "Listado de autores",
        "description": "",
        "operationId": "listAuthor",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Author"
        ],
        "summary": "Add a new Author in the register",
        "description": "",
        "operationId": "addAuthor",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Author object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Author"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Author created"
          }
        }
      },
      "put": {
        "tags": [
          "Author"
        ],
        "summary": "Update an existing Author",
        "description": "",
        "operationId": "updateAuthor",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Author object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Author"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Author not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/Author/{id}": {
      "get": {
        "tags": [
          "Author"
        ],
        "summary": "Find loan by ID",
        "description": "Returns a single loan",
        "operationId": "getAuthorbyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Author return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Author"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Author not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Author"
        ],
        "summary": "Updates a student in the register with form data",
        "description": "",
        "operationId": "updateAuthorWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Author that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the pet",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the pet",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "Author"
        ],
        "summary": "Deletes a Student",
        "description": "",
        "operationId": "deleteAuthor",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Author id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Author not found"
          }
        }
      }
    },
    "/Editorial": {
      "get": {
        "tags": [
          "Editorial"
        ],
        "summary": "Listado de editoriales",
        "description": "",
        "operationId": "listEditorial",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "List"
          }
        }
      },
      "post": {
        "tags": [
          "Editorial"
        ],
        "summary": "Add a new Editorial in the register",
        "description": "",
        "operationId": "addEditorial",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Editorial object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Editorial"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Author created"
          }
        }
      },
      "put": {
        "tags": [
          "Editorial"
        ],
        "summary": "Update an existing Editorial",
        "description": "",
        "operationId": "updateEditorial",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Editorial object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Editorial"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Editorial not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/Editorial/{id}": {
      "get": {
        "tags": [
          "Editorial"
        ],
        "summary": "Find Editorial by ID",
        "description": "Returns a single loan",
        "operationId": "getEditorialbyid",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Editorial return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Editorial"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Editorial not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Editorial"
        ],
        "summary": "Updates a Editorial in the register with form data",
        "description": "",
        "operationId": "updateEditorialWithForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Editorial that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated name of the Editorial",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Updated status of the Editorial",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "Editorial"
        ],
        "summary": "Deletes a Editorial",
        "description": "",
        "operationId": "deleteEditorial",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Editorial id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Editorial not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Book": {
      "type": "object",
      "required": [
        "title",
        "stateBook",
        "commentsBook"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "title": {
          "type": "string",
          "example": "Juan Quijote de la mancha"
        },
        "stateBook": {
          "type": "string"
        },
        "commentsBook": {
          "type": "integer"
        }
      }
    },
...
    "Author": {
      "type": "object",
      "required": [
        "id_author",
        "name_Author",
        "country"
      ],
      "properties": {
        "id_author": {
          "type": "string",
          "format": "uuid"
        },
        "name_Author": {
          "type": "string",
          "example": "Ensias sangrantes"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "Editorial": {
      "type": "object",
      "required": [
        "id_editorial",
        "name_editorial",
        "direction",
        "contact_phone"
      ],
      "properties": {
        "id_editorial": {
          "type": "string",
          "format": "uuid"
        },
        "name_editorial": {
          "type": "string",
          "example": "Editorial PS"
        },
        "direction": {
          "type": "string"
        },
        "contact_phone": {
          "type": "string"
        }
      }
    },
...
  }
}